[OrderBundle, WorkflowBundle] refactor state change logging for workflows -> move classes to workflow bundle
