+++ src/CoreShop/Bundle/OrderBundle/Resources/config/services/workflow.yml
@@ -2,7 +2,6 @@ parameters:
     _defaults:
         public: true
 
-    coreshop.workflow.manager.order.history_log_state_identifier: !php/const CoreShop\Component\Order\Notes::ORDER_HISTORY_STATE_LOG
     coreshop.workflow.include_invoice_state_to_complete_order: false
 
 services:
@@ -100,24 +99,15 @@ services:
     coreshop.workflow.state_manager:
         class: CoreShop\Bundle\OrderBundle\Workflow\WorkflowStateManager
         arguments:
-            - '@event_dispatcher'
-            - '@coreshop.state_machine_manager'
-            - '@pimcore.translator'
-            - '%coreshop.workflow.manager.order.history_log_state_identifier%'
-            - '%coreshop.state_machine.colors%'
+            - '@coreshop.workflow.state_info_manager'
 
     coreshop.workflow.order_history_logger:
         class: CoreShop\Bundle\OrderBundle\Workflow\OrderHistoryLogger
         arguments:
             - '@coreshop.repository.order'
-            - '@coreshop.object_note_service'
-            - '@pimcore.translator'
-            - '%coreshop.workflow.manager.order.history_log_state_identifier%'
+            - '@coreshop.workflow.history_logger'
 
     coreshop.workflow.order_state_history_logger:
         class: CoreShop\Bundle\OrderBundle\Workflow\OrderStateHistoryLogger
         arguments:
-            - '@coreshop.state_machine_manager'
-            - '@coreshop.object_note_service'
-            - '@pimcore.translator'
-            - '%coreshop.workflow.manager.order.history_log_state_identifier%'
+            - '@coreshop.workflow.state_history_logger'
