+++ src/CoreShop/Bundle/OrderBundle/Resources/translations/admin.de.yml
@@ -209,7 +209,6 @@ coreshop_workflow_transition_coreshop_shipment_cancel: 'Stornieren'
 coreshop_workflow_transition_coreshop_invoice_create: 'Erstellen'
 coreshop_workflow_transition_coreshop_invoice_complete: 'Abschließen'
 coreshop_workflow_transition_coreshop_invoice_cancel: 'Stornieren'
-coreshop_workflow_order_history_logger_prefix: 'Log'
 coreshop_discount: 'Rabatt (inkl. Mwst.)'
 coreshop_discount_tax: 'Rabatt Mwst.'
 coreshop_discount_without_tax: 'Rabatt (exkl. Mwst.)'
