+++ src/CoreShop/Bundle/OrderBundle/Resources/translations/admin.en.yml
@@ -208,7 +208,6 @@ coreshop_workflow_transition_coreshop_shipment_cancel: 'Cancel'
 coreshop_workflow_transition_coreshop_invoice_create: 'Create'
 coreshop_workflow_transition_coreshop_invoice_complete: 'Complete'
 coreshop_workflow_transition_coreshop_invoice_cancel: 'Cancel'
-coreshop_workflow_order_history_logger_prefix: 'Log'
 coreshop_discount: 'Discount (incl. Tax)'
 coreshop_discount_tax: 'Discount Tax'
 coreshop_discount_without_tax: 'Discount (excl. Tax.)'
