+++ src/CoreShop/Bundle/WorkflowBundle/Resources/config/services/workflow.yml
@@ -1,3 +1,9 @@
+parameters:
+    _defaults:
+        public: true
+
+    coreshop.workflow.manager.history_log_state_identifier: !php/const CoreShop\Bundle\WorkflowBundle\Notes::HISTORY_STATE_LOG
+
 services:
     _defaults:
         public: true
@@ -26,3 +32,31 @@ services:
             - '@service_container'
         tags:
             - { name: kernel.event_subscriber }
+
+    coreshop.workflow.history_logger:
+        class: CoreShop\Bundle\WorkflowBundle\History\HistoryLogger
+        arguments:
+            - '@coreshop.object_note_service'
+            - '@Pimcore\Translation\Translator'
+            - '%coreshop.workflow.manager.history_log_state_identifier%'
+
+    coreshop.workflow.state_history_logger:
+        class: CoreShop\Bundle\WorkflowBundle\History\StateHistoryLogger
+        arguments:
+            - '@coreshop.object_note_service'
+            - '@Pimcore\Translation\Translator'
+            - '%coreshop.workflow.manager.history_log_state_identifier%'
+
+    coreshop.repository.workflow_history:
+        class: CoreShop\Bundle\WorkflowBundle\History\HistoryRepository
+        arguments:
+            - '%coreshop.workflow.manager.history_log_state_identifier%'
+
+    coreshop.workflow.state_info_manager:
+        class: CoreShop\Bundle\WorkflowBundle\StateManager\WorkflowStateInfoManager
+        arguments:
+            - '@event_dispatcher'
+            - '@coreshop.state_machine_manager'
+            - '@translator'
+            - '@coreshop.repository.workflow_history'
+            - '%coreshop.state_machine.colors%'
