+++ src/CoreShop/Bundle/WorkflowBundle/Resources/config/translations/admin.de.yml
@@ -1,0 +1,2 @@
+---
+coreshop_workflow_history_logger_prefix: 'Log'
