+++ src/CoreShop/Bundle/WorkflowBundle/Resources/config/translations/admin.en.yml
@@ -1,0 +1,1 @@
+coreshop_workflow_history_logger_prefix: 'Log'
